feat: switch to `wasm-opt` cli
