+++ src/web/mod.rs
@@ -6,5 +6,3 @@ pub(crate) mod bundle;
 pub(crate) mod profiles;
 pub(crate) mod run;
 pub(crate) mod serve;
-#[cfg(feature = "wasm-opt")]
-pub(crate) mod wasm_opt;
